+++ CHANGES.md
@@ -9,6 +9,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 - Rename library from "ninepatch" to "quilt"
 
+### Fixed
+
+- Potential issue with require path
+
 ## [1.0.0] - 2023-11-03
 
 Initial release